# Optimize String visitor by hand