@byroot 2025-02-13
Oh, I just realized we're doing all this because of https://github.com/rails/rails/pull/47129

Feels like we could just be doing it lazily.